+++ lib/components/widgets/appbar/root_app_bar.dart
@@ -1,13 +1,14 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:root/components/root_images.dart';
+import 'package:root/presentation/notification/check_notification_screen.dart';
 
-import '../root_colors.dart';
+import '../../root_colors.dart';
 
-class RootDetailAppBar extends StatelessWidget implements PreferredSizeWidget {
+class RootAppBar extends StatelessWidget implements PreferredSizeWidget {
   final String title;
 
-  const RootDetailAppBar({
+  const RootAppBar({
     required this.title,
     super.key,
   });
@@ -20,16 +21,6 @@ class RootDetailAppBar extends StatelessWidget implements PreferredSizeWidget {
         padding: EdgeInsets.symmetric(horizontal: 10.0.w),
         child: AppBar(
           automaticallyImplyLeading: false,
-          leading: IconButton(
-            onPressed: () {
-              Navigator.of(context).pop(context);
-            },
-            icon: Image.asset(
-              RootImages.backIcon,
-              width: 25.w,
-              height: 25.h,
-            ),
-          ),
           title: Text(
             title,
             style: TextStyle(
@@ -38,6 +29,22 @@ class RootDetailAppBar extends StatelessWidget implements PreferredSizeWidget {
               fontWeight: FontWeight.w600,
             ),
           ),
+          actions: [
+            IconButton(
+              onPressed: () {
+                Navigator.of(context).push(
+                  MaterialPageRoute(
+                    builder: (_) => const CheckNotificationScreen(),
+                  ),
+                );
+              },
+              icon: Image.asset(
+                RootImages.bell,
+                width: 25.w,
+                height: 25.h,
+              ),
+            ),
+          ],
           centerTitle: true,
         ),
       ),
@@ -45,5 +52,5 @@ class RootDetailAppBar extends StatelessWidget implements PreferredSizeWidget {
   }
 
   @override
-  Size get preferredSize => Size.fromHeight(45.h);
+  Size get preferredSize => Size.fromHeight(50.h);
 }
