+++ lib/components/widgets/appbar/root_detail_app_bar.dart
@@ -1,14 +1,13 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:root/components/root_images.dart';
-import 'package:root/presentation/notification/notification_screen.dart';
 
-import '../root_colors.dart';
+import '../../root_colors.dart';
 
-class RootAppBar extends StatelessWidget implements PreferredSizeWidget {
+class RootDetailAppBar extends StatelessWidget implements PreferredSizeWidget {
   final String title;
 
-  const RootAppBar({
+  const RootDetailAppBar({
     required this.title,
     super.key,
   });
@@ -21,6 +20,16 @@ class RootAppBar extends StatelessWidget implements PreferredSizeWidget {
         padding: EdgeInsets.symmetric(horizontal: 10.0.w),
         child: AppBar(
           automaticallyImplyLeading: false,
+          leading: IconButton(
+            onPressed: () {
+              Navigator.of(context).pop(context);
+            },
+            icon: Image.asset(
+              RootImages.backIcon,
+              width: 25.w,
+              height: 25.h,
+            ),
+          ),
           title: Text(
             title,
             style: TextStyle(
@@ -29,22 +38,6 @@ class RootAppBar extends StatelessWidget implements PreferredSizeWidget {
               fontWeight: FontWeight.w600,
             ),
           ),
-          actions: [
-            IconButton(
-              onPressed: () {
-                Navigator.of(context).push(
-                  MaterialPageRoute(
-                    builder: (_) => const NotificationScreen(),
-                  ),
-                );
-              },
-              icon: Image.asset(
-                RootImages.bell,
-                width: 25.w,
-                height: 25.h,
-              ),
-            ),
-          ],
           centerTitle: true,
         ),
       ),
@@ -52,5 +45,5 @@ class RootAppBar extends StatelessWidget implements PreferredSizeWidget {
   }
 
   @override
-  Size get preferredSize => Size.fromHeight(50.h);
+  Size get preferredSize => Size.fromHeight(45.h);
 }
