+++ lib/components/widgets/appbar/root_main_app_bar.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:root/components/root_images.dart';
-import 'package:root/presentation/notification/notification_screen.dart';
+import 'package:root/presentation/notification/check_notification_screen.dart';
 
 class RootMainAppBar extends StatelessWidget implements PreferredSizeWidget {
   const RootMainAppBar({super.key,});
@@ -26,7 +26,7 @@ class RootMainAppBar extends StatelessWidget implements PreferredSizeWidget {
                 onPressed: () {
                   Navigator.of(context).push(
                     MaterialPageRoute(
-                      builder: (_) => const NotificationScreen(),
+                      builder: (_) => const CheckNotificationScreen(),
                     ),
                   );
                 },
